refactor(bin-designer): split migration logic out of constants/defaults
defaults.ts had grown to 1020 lines, three quarters of it backward-compat
migration control flow interleaved with the DEFAULT_* data it backfills.
All migration now lives in paramMigration.ts (which already held
migrateWalls and was documented as the migration home):

- lid shims (clickRails/coverage/extraHeight/attachment/retention/tray),
  the color-era migrations (legacy slot colors, lip grid eras, topAccent,
  migrateFeatureColors), migrateParams + MigrateParamsInput, and
  extractStyleDefaults/STYLE_DEFAULT_OMIT_KEYS
- defaults.ts is now 257 lines of pure data importing only types; the
  import direction is strictly paramMigration -> defaults (no cycle),
  and its previous max-lines warning is gone
- VALID_WALL_PATTERNS/VALID_HANDLE_SHAPES and the DEFAULT_* configs the
  migration backfills from are now exported from defaults
- constants barrel re-exports migration API from its new home; the six
  deep importers of migrateParams point at paramMigration directly
- defaults.test.ts split the same way: 82 migration tests moved to
  paramMigration.test.ts (118/1094 lines after)
